bot/requirements.txt
aiogram~=3.21.0
python-decouple~=3.8
asyncio~=3.4.3
aiosqlite~=0.21.0